test_controller/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
pytz==2020.1
gunicorn==19.10.0
whitenoise==6.0.0
django-crispy-forms==1.14.0